flask/requirements.txt
mysql-connector-python
PyYAML
requests
flask-socketio
flask
schedule